+++ src/Client.php
@@ -325,10 +325,15 @@ class Client implements ClientInterface, \Psr\Http\Client\ClientInterface
     {
         $request = $this->applyOptions($request, $options);
 
-        if ('' === $request->getProtocolVersion()) {
+        $protocolVersion = $request->getProtocolVersion();
+
+        if ('' === $protocolVersion) {
             \trigger_deprecation('guzzlehttp/guzzle', '7.11', 'Sending a request with an empty protocol version is deprecated; guzzlehttp/guzzle 8.0 will reject empty protocol versions.');
 
             $request = Psr7\Utils::modifyRequest($request, ['version' => '1.1']);
+        } elseif (!self::isProtocolVersionValid($protocolVersion)) {
+            \trigger_deprecation('guzzlehttp/guzzle', '7.11', 'Sending a request with a malformed protocol version is deprecated; guzzlehttp/guzzle 8.0 will reject malformed protocol versions.');
+            $options['_guzzle_malformed_protocol_version_deprecated'] = true;
         }
 
         /** @var HandlerStack $handler */
@@ -486,6 +491,11 @@ class Client implements ClientInterface, \Psr\Http\Client\ClientInterface
         }
 
         return \is_float($version) ? \number_format($version, 1, '.', '') : (string) $version;
+    }
+
+    private static function isProtocolVersionValid(string $version): bool
+    {
+        return 1 === \preg_match('/^\d+(?:\.\d+)?$/D', $version);
     }
 
     /**